requirements.txt
absl-py==1.4.0
altgraph==0.17.3
ao3-api==2.3.0
appdirs==1.4.4
argcomplete==2.0.0
astunparse==1.6.3
async-generator==1.10
attrs==23.1.0
Babel==2.11.0
bcrypt==3.2.2
Beaker==1.12.1
beautifulsoup4==4.12.2
blivet==3.7.1
blivet-gui==2.4.2
Brlapi==0.8.4
cached-property==1.5.2
cachetools==5.3.0
certifi==2022.9.24
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
cryptography==37.0.2
cson==0.8
cssselect==1.1.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
decorator==5.1.1
distro==1.8.0
docker==5.0.3
docker-compose==1.29.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.1.1
fedora-third-party==0.10
file-magic==0.4.0
Flask==2.2.3
flatbuffers==23.1.21
fros==1.1
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpg==1.17.1
grpcio==1.51.3
gssapi==1.7.3
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.8.0
humanize==3.13.1
idna==3.4
imageio==2.31.3
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
jax==0.4.4
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.12.0rc0
keyboard==0.13.5
koji==1.33.0
langtable==0.0.63
libclang==15.0.6.1
libcomps==0.1.18
lxml==4.9.2
Mako==1.2.3
Markdown==3.4.1
MarkupSafe==2.1.2
MouseInfo==0.1.3
msmcauth==0.0.4
nftables==0.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
olefile==0.46
opencv-python==4.7.0.68
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
paramiko==3.1.0
Paste==3.5.2
pexpect==4.8.0
pickle-mixin==1.0.2
pid==2.2.3
Pillow==9.4.0
ply==3.11
productmd==1.36
progressbar2==3.53.2
protobuf==4.22.0
psutil==5.9.5
ptyprocess==0.7.0
pwquality==1.4.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycairo==1.23.0
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.2
pyenchant==3.2.2
PyGetWindow==0.0.9
pygit2==1.12.2
Pygments==2.14.0
PyGObject==3.44.1
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
pykickstart==3.47
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparted==3.12.0
pyperclip==1.8.2
PyQt5==5.15.9
PyQt5-sip==12.11.1
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
PySimpleGUI==4.60.5
PySocks==1.7.1
pytesseract==0.3.10
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-meh==0.51
python-utils==3.3.3
python3-xlib==0.15
pytweening==1.0.4
pytz==2023.3.post1
pyudev==0.24.0
pyvips==2.2.1
pyxdg==0.27
PyYAML==6.0
regex==2023.6.3
requests==2.28.2
requests-file==1.5.1
requests-ftp==0.3.1
requests-gssapi==1.2.3
requests-oauthlib==1.3.1
rpm==4.18.1
rpmautospec==0.3.5
rpmconf==1.1.8
rsa==4.9
scipy==1.10.1
scour==0.38.2
selenium==4.8.3
selinux==3.5
sepolicy==3.5
setools==4.4.3
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sos==4.5.1
soupsieve==2.4.1
speg==0.3
systemd-python==235
Tempita==0.5.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0rc0
tensorflow-estimator==2.12.0rc0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
texttable==1.6.4
torch==1.13.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.16
uuid==1.30
webptools==0.0.9
websocket-client==1.3.3
Werkzeug==2.2.3
wrapt==1.14.1
wsproto==1.2.0
